Adds proguard to prevent Mediapipe from minifying.
Minifying brakes Mediapipe's protobuf, to fix this, we need to tell gradle to make an exemption. We tell it not to minify anything in the proguard-rules.pro file, and then put Mediapipe's protobuf as an exception. This is done in the example. As we cannot do this automatically for users in the library, we specify this in the README so users can replicate the results.